refactor: move test utilities into tests/utils
