+++ openfold3/tests/test_embedders.py
@@ -28,7 +28,7 @@ from openfold3.core.model.latent.msa_module import MSAModuleStack
 from openfold3.core.model.latent.template_module import TemplateEmbedderAllAtom
 from openfold3.projects.of3_all_atom.project_entry import OF3ProjectEntry
 from openfold3.tests.config import consts
-from openfold3.tests.data_utils import random_asym_ids, random_of3_features
+from openfold3.tests.utils.data_utils import random_asym_ids, random_of3_features
 
 
 class TestInputEmbedderAllAtom:
