+++ openfold3/tests/test_heads.py
@@ -31,8 +31,8 @@ from openfold3.projects.of3_all_atom.config.model_config import (
 )
 from openfold3.projects.of3_all_atom.project_entry import OF3ProjectEntry
 from openfold3.tests.config import consts
-from openfold3.tests.data_utils import random_of3_features
-from openfold3.tests.model_utils import initialize_model_weights
+from openfold3.tests.utils.data_utils import random_of3_features
+from openfold3.tests.utils.model_utils import initialize_model_weights
 
 
 class TestPredictedAlignedErrorHead(unittest.TestCase):
