+++ openfold3/tests/test_of3_model.py
@@ -20,9 +20,9 @@ from openfold3.core.utils.precision_utils import OF3DeepSpeedPrecision
 from openfold3.core.utils.tensor_utils import tensor_tree_map
 from openfold3.projects.of3_all_atom.project_entry import OF3ProjectEntry
 from openfold3.projects.of3_all_atom.runner import OpenFold3AllAtom
-from openfold3.tests import compare_utils
 from openfold3.tests.config import consts
-from openfold3.tests.data_utils import random_of3_features
+from openfold3.tests.utils import compare_utils
+from openfold3.tests.utils.data_utils import random_of3_features
 
 
 class TestOF3Model:
