+++ openfold3/tests/test_sequence_local_atom_attention.py
@@ -26,7 +26,7 @@ from openfold3.core.model.layers.sequence_local_atom_attention import (
 )
 from openfold3.core.utils.tensor_utils import tensor_tree_map
 from openfold3.tests.config import consts
-from openfold3.tests.data_utils import random_of3_features
+from openfold3.tests.utils.data_utils import random_of3_features
 
 C_ATOM_REF = mlc.ConfigDict(
     {
